add some tests

+++ queue/priority_blocking_queue.go
@@ -11,7 +11,7 @@ import (
 )
 
 // NewPriorityBlockingQueue new priority blocking queue
-func NewPriorityBlockingQueue[E any](comparator contract.Comparator[E], cap int) *PriorityBlockingQueue[E] {
+func NewPriorityBlockingQueue[E any](comparator contract.Comparator[E], cap int64) *PriorityBlockingQueue[E] {
 	queue := new(PriorityBlockingQueue[E])
 	queue.items = NewPriorityQueue(comparator)
 	queue.takeLock = sync.NewCond(queue.items)
@@ -23,11 +23,12 @@ func NewPriorityBlockingQueue[E any](comparator contract.Comparator[E], cap int)
 // PriorityBlockingQueue priority blocking queue
 type PriorityBlockingQueue[E any] struct {
 	items    *PriorityQueue[E]
-	cap      int
+	cap      int64
 	takeLock *sync.Cond
 	putLock  *sync.Cond
 }
 
+// Count returns the size of queue
 func (q *PriorityBlockingQueue[E]) Count() int64 {
 	if q.items.TryRLock() {
 		defer q.items.RUnlock()
@@ -35,6 +36,7 @@ func (q *PriorityBlockingQueue[E]) Count() int64 {
 	return q.items.Count()
 }
 
+// IsEmpty returns whether the queue is empty
 func (q *PriorityBlockingQueue[E]) IsEmpty() bool {
 	if q.items.TryRLock() {
 		defer q.items.RUnlock()
@@ -42,6 +44,7 @@ func (q *PriorityBlockingQueue[E]) IsEmpty() bool {
 	return q.items.IsEmpty()
 }
 
+// IsNotEmpty returns whether the queue is not empty
 func (q *PriorityBlockingQueue[E]) IsNotEmpty() bool {
 	if q.items.TryRLock() {
 		defer q.items.RUnlock()
@@ -49,6 +52,7 @@ func (q *PriorityBlockingQueue[E]) IsNotEmpty() bool {
 	return q.items.IsNotEmpty()
 }
 
+// Clear clears the queue
 func (q *PriorityBlockingQueue[E]) Clear() {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
@@ -56,6 +60,7 @@ func (q *PriorityBlockingQueue[E]) Clear() {
 	q.items.Clear()
 }
 
+// Peek returns the first element of the queue
 func (q *PriorityBlockingQueue[E]) Peek() (E, bool) {
 	if q.items.TryRLock() {
 		defer q.items.RUnlock()
@@ -63,11 +68,12 @@ func (q *PriorityBlockingQueue[E]) Peek() (E, bool) {
 	return q.items.Peek()
 }
 
+// TryEnqueue enqueues a new element into the queue, it will return false if the size is up to the capacity
 func (q *PriorityBlockingQueue[E]) TryEnqueue(value E) bool {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
 	}
-	if q.cap == int(q.items.Count()) {
+	if q.cap == q.items.Count() {
 		return false
 	}
 	ok := q.items.Enqueue(value)
@@ -75,6 +81,8 @@ func (q *PriorityBlockingQueue[E]) TryEnqueue(value E) bool {
 	return ok
 }
 
+// TryDequeue dequeues the first element of the queue and returns it.
+// The empty value of the element type and false will be returned when the queue is empty
 func (q *PriorityBlockingQueue[E]) TryDequeue() (E, bool) {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
@@ -87,11 +95,12 @@ func (q *PriorityBlockingQueue[E]) TryDequeue() (E, bool) {
 	return value, ok
 }
 
+// Enqueue enqueues a new element into the queue, it will block if the size is up to capacity
 func (q *PriorityBlockingQueue[E]) Enqueue(value E) bool {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
 	}
-	for q.cap == int(q.items.Count()) {
+	for q.cap == q.items.Count() {
 		q.putLock.Wait()
 	}
 	ok := q.items.Enqueue(value)
@@ -99,6 +108,7 @@ func (q *PriorityBlockingQueue[E]) Enqueue(value E) bool {
 	return ok
 }
 
+// Dequeue dequeues the first element of queue, it will block if the queue is empty
 func (q *PriorityBlockingQueue[E]) Dequeue() (E, bool) {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
@@ -111,6 +121,9 @@ func (q *PriorityBlockingQueue[E]) Dequeue() (E, bool) {
 	return value, ok
 }
 
+// EnqueueTimeout enqueues element into the queue.
+// It will block when the size of queue is up to capacity.
+// It will return true if the element is successfully enqueued or false when time is out
 func (q *PriorityBlockingQueue[E]) EnqueueTimeout(value E, duration time.Duration) bool {
 	timeout := time.After(duration)
 	done := make(chan struct{})
@@ -132,6 +145,9 @@ func (q *PriorityBlockingQueue[E]) EnqueueTimeout(value E, duration time.Duratio
 	}
 }
 
+// DequeueTimeout removes the first element and returns it.
+// It will block when the queue is empty.
+// It will return zero value and false when time is out
 func (q *PriorityBlockingQueue[E]) DequeueTimeout(duration time.Duration) (value E, ok bool) {
 	timeout := time.After(duration)
 	done := make(chan struct{})
@@ -153,6 +169,7 @@ func (q *PriorityBlockingQueue[E]) DequeueTimeout(duration time.Duration) (value
 	}
 }
 
+// Remove removes the specific element
 func (q *PriorityBlockingQueue[E]) Remove(value E) {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
@@ -160,6 +177,7 @@ func (q *PriorityBlockingQueue[E]) Remove(value E) {
 	q.items.Remove(value)
 }
 
+// RemoveWhere removes elements which matches the callback
 func (q *PriorityBlockingQueue[E]) RemoveWhere(callback func(E) bool) {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
@@ -167,6 +185,7 @@ func (q *PriorityBlockingQueue[E]) RemoveWhere(callback func(E) bool) {
 	q.items.RemoveWhere(callback)
 }
 
+// ToArray converts to array
 func (q *PriorityBlockingQueue[E]) ToArray() []E {
 	if q.items.TryRLock() {
 		defer q.items.RUnlock()
@@ -174,6 +193,7 @@ func (q *PriorityBlockingQueue[E]) ToArray() []E {
 	return q.items.ToArray()
 }
 
+// ToJSON converts to json
 func (q *PriorityBlockingQueue[E]) ToJSON() ([]byte, error) {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
@@ -181,10 +201,12 @@ func (q *PriorityBlockingQueue[E]) ToJSON() ([]byte, error) {
 	return q.items.ToJSON()
 }
 
+// MarshalJSON implements [json.Marshaller]
 func (q *PriorityBlockingQueue[E]) MarshalJSON() ([]byte, error) {
 	return q.ToJSON()
 }
 
+// UnmarshalJSON implements [json.Unmarshaller]
 func (q *PriorityBlockingQueue[E]) UnmarshalJSON(data []byte) error {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
@@ -195,7 +217,7 @@ func (q *PriorityBlockingQueue[E]) UnmarshalJSON(data []byte) error {
 	}
 	q.items.Clear()
 	for _, value := range values {
-		for q.cap == int(q.items.Count()) {
+		for q.cap == q.items.Count() {
 			q.putLock.Wait()
 		}
 		q.items.Enqueue(value)
@@ -204,6 +226,7 @@ func (q *PriorityBlockingQueue[E]) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// String converts to string
 func (q *PriorityBlockingQueue[E]) String() string {
 	if q.items.TryLock() {
 		defer q.items.Unlock()
